fix: cast video tensor to uint8 before write_video
Fixes TypeError with newer PyAV versions where pict_type
requires an integer. write_video expects uint8 tensors.

+++ inference.py
@@ -173,6 +173,6 @@ for i, batch_data in tqdm(enumerate(dataloader), disable=(local_rank != 0)):
         output_path = os.path.join(args.output_folder, f'video_{args.start_index + i:03d}.mp4')
     else:
         output_path = os.path.join(args.output_folder, f'{prompt[:100]}.mp4')
-    write_video(output_path, video[0], fps=16)
+    write_video(output_path, video[0].to(torch.uint8), fps=16)
